+++ dbt/include/mysql5/macros/adapters.sql
@@ -1,5 +1,5 @@
 
-{% macro mysql__list_schemas(database) %}
+{% macro mysql5__list_schemas(database) %}
     {% call statement('list_schemas', fetch_result=True, auto_begin=False) -%}
         select distinct schema_name
         from information_schema.schemata
@@ -8,47 +8,45 @@
     {{ return(load_result('list_schemas').table) }}
 {% endmacro %}
 
-{% macro mysql__create_schema(relation) -%}
+{% macro mysql5__create_schema(relation) -%}
   {%- call statement('create_schema') -%}
     create schema if not exists {{ relation.without_identifier() }}
   {%- endcall -%}
 {% endmacro %}
 
-{% macro mysql__drop_schema(relation) -%}
+{% macro mysql5__drop_schema(relation) -%}
   {%- call statement('drop_schema') -%}
     drop schema if exists {{ relation.without_identifier() }}
   {% endcall %}
 {% endmacro %}
 
-{% macro mysql__drop_relation(relation) -%}
+{% macro mysql5__drop_relation(relation) -%}
     {% call statement('drop_relation', auto_begin=False) -%}
         drop {{ relation.type }} if exists {{ relation }}
     {%- endcall %}
 {% endmacro %}
 
-{% macro mysql__truncate_relation(relation) -%}
+{% macro mysql5__truncate_relation(relation) -%}
     {% call statement('truncate_relation') -%}
       truncate table {{ relation }}
     {%- endcall %}
 {% endmacro %}
 
-{% macro mysql__create_table_as(temporary, relation, sql) -%}
+{% macro mysql5__create_table_as(temporary, relation, sql) -%}
   {%- set sql_header = config.get('sql_header', none) -%}
 
   {{ sql_header if sql_header is not none }}
 
   create {% if temporary: -%}temporary{%- endif %} table
     {{ relation.include(database=False) }}
-  as (
     {{ sql }}
-  )
 {% endmacro %}
 
-{% macro mysql__current_timestamp() -%}
+{% macro mysql5__current_timestamp() -%}
   current_timestamp()
 {%- endmacro %}
 
-{% macro mysql__rename_relation(from_relation, to_relation) -%}
+{% macro mysql5__rename_relation(from_relation, to_relation) -%}
   {#
     MySQL rename fails when the relation already exists, so a 2-step process is needed:
     1. Drop the existing relation
@@ -62,12 +60,12 @@
   {% endcall %}
 {% endmacro %}
 
-{% macro mysql__check_schema_exists(database, schema) -%}
+{% macro mysql5__check_schema_exists(database, schema) -%}
     {# no-op #}
     {# see MySQLAdapter.check_schema_exists() #}
 {% endmacro %}
 
-{% macro mysql__get_columns_in_relation(relation) -%}
+{% macro mysql5__get_columns_in_relation(relation) -%}
     {% call statement('get_columns_in_relation', fetch_result=True) %}
         show columns from {{ relation.schema }}.{{ relation.identifier }}
     {% endcall %}
@@ -76,7 +74,7 @@
     {{ return(sql_convert_columns_in_relation(table)) }}
 {% endmacro %}
 
-{% macro mysql__list_relations_without_caching(schema_relation) %}
+{% macro mysql5__list_relations_without_caching(schema_relation) %}
   {% call statement('list_relations_without_caching', fetch_result=True) -%}
     select
       null as "database",
@@ -92,6 +90,6 @@
   {{ return(load_result('list_relations_without_caching').table) }}
 {% endmacro %}
 
-{% macro mysql__generate_database_name(custom_database_name=none, node=none) -%}
+{% macro mysql5__generate_database_name(custom_database_name=none, node=none) -%}
   {% do return(None) %}
 {%- endmacro %}